utilities/supporting-stacks/github-creds/requirements.txt
aws-cdk-lib==2.151.0
constructs>=10.0.0,<11.0.0
pydantic==1.10.17
../../packages/python/provena-shared-functionality
